Separate -l/--limit into -l/--search-limit & -L/--results-limit
 The default search limit is 30. The default results limit is 10.

 So, we still show 10 results by default. But the results are
 the top 10 from 30 ranked.

 This carries the risk of pushing the most relevant result(s) out of
 the top 10. But with the searches I've tried, the new behavior seems
 to always provide better (or the same) results.

 Bump version to 0.2 while at it.

+++ src/esr_printer.rs
@@ -138,7 +138,7 @@ impl EsrPrinter {
         println!("{}", self.red_bold(&msg));
     }
 
-    pub fn limit_out_of_range(&self, limit: u8, min: u8, max: u8) {
+    pub fn limit_out_of_range(&self, limit: usize, min: usize, max: usize) {
         let msg = format!("{} is out of the range of valid limits. \
                           Please pass a value between {} and {}.", limit, min, max);
         println!("{}", self.yellow_bold(&msg));
